reorganize css and framework code into lib and style dirs
