+++ app/renderer/index.html
@@ -5,7 +5,7 @@
     <title>Appium</title>
     <link rel="stylesheet" href="../../node_modules/font-awesome/css/font-awesome.min.css" />
     <link rel="stylesheet" href="../../node_modules/antd/dist/antd.css" />
-    <link rel="stylesheet" href="./highlighting/github-gist.css" />
+    <link rel="stylesheet" href="./styles/github-gist.css" />
     <script>
       (function() {
         if (!process.env.HOT) {
